requirements.txt
requests>=2.31.0
inquirer>=3.1.3
pyyaml>=6.0.1